mcp-client-server/pyproject.toml
[project]
name = "mcp-client-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.0.10",
    "langgraph>=0.4.2",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
]
